Branding update, the official icons are now used for the clients.
Previously I was skeptical about using the official iconography over concerns about trademark and quality of the snap.
After nearly a year of use, the quality of the snap has proven to be high with plenty of feedback.

The trademark concerns are still legitimate, however, UK law applies in this instance, with Canonical, Jagex, and myself being bound by English law.
As per Section 11, (2)(c) of the Trademark Act 1994, this package is being provided as an accessory for the official client, providing compatability and security benefits.
As a result, and keeping in mind the trademarks are being used in honest faith with no personal profiteering, in a package held to high quality, the official icons are now used.

Embedding the client executable code itself directly still cannot be done due to copyright law and an explicit license forbidding it.